refactor indentation code to seperate module
